test: modularize tps cli test suite
